DOC: refactored custom forces, starting conformations into contrib
