+++ contrib/forces.py
@@ -1,0 +1,63 @@
+r"""
+Array of spherical wells
+========================
+
+Confine each subchain in simulation to a sherical well.
+Should be used in combination with contrib.initialize_chains.
+
+"""
+
+try:
+    import openmm
+except Exception:
+    import simtk.openmm as openmm
+
+
+def spherical_well_array(sim_object, r, cell_size, particles=None,
+                         width=1, depth=1, name="spherical_well_array"):
+    """
+    An (array of) spherical potential wells. Uses floor functions to map
+    particle positions to the coordinates of the well.
+
+    Parameters
+    ----------
+
+    r : float
+        Radius of the nucleus
+    cell_size : float
+        width of cell in lattice of spherical wells
+    particles : list of int or np.array
+        indices of particles that are attracted
+    width : float, optional
+        Width of attractive well, nm.
+    depth : float, optional
+        Depth of attractive potential in kT
+        Positive means the walls are repulsive (i.e chain confined within lamina).
+        Negative means walls are attractive (i.e. attraction to lamina)
+    """
+
+    force = openmm.CustomExternalForce(
+        "step(1+d) * step(1-d) * SPHWELLdepth * (1 + cos(3.1415926536*d)) / 2;"
+        "d = (sqrt((x1-SPHWELLx)^2 + (y1-SPHWELLy)^2 + (z1-SPHWELLz)^2) - SPHWELLradius) / SPHWELLwidth;"
+        "x1 = x - L*floor(x/L);"
+        "y1 = y - L*floor(y/L);"
+        "z1 = z - L*floor(z/L);"
+    )
+    force.name = name
+    particles = range(sim_object.N) if particles is None else particles
+    center = 3 * [cell_size/2]
+    
+    force.addGlobalParameter("SPHWELLradius", r * sim_object.conlen)
+    force.addGlobalParameter("SPHWELLwidth", width * sim_object.conlen)
+    force.addGlobalParameter("SPHWELLdepth", depth * sim_object.kT)
+    force.addGlobalParameter("L", cell_size * sim_object.conlen)
+    force.addGlobalParameter("SPHWELLx", center[0] * sim_object.conlen)
+    force.addGlobalParameter("SPHWELLy", center[1] * sim_object.conlen)
+    force.addGlobalParameter("SPHWELLz", center[2] * sim_object.conlen)
+
+    # adding all the particles on which force acts
+    for i in particles:
+        # NOTE: the explicit type cast seems to be necessary if we have an np.array...
+        force.addParticle(int(i), [])
+
+    return force
